+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added `Small/Core.hpp`, `Small/Graphics.hpp` and `Small/System.hpp` to include all the files from the specific folders
 - the tilemap can be passed a new texture using `void Tilemap::use(sf::Texture*)`
 - adding github actions to build the project after each commit/pull request
+- added a `sgl::Graphics::TileView` to use tiles coordinates to place a view
 
 ### Changed
 - the tilemap has been updated to no longer hold the texture, and be able to modify the vertex array when needed